+++ pephub/const.py
@@ -35,3 +35,5 @@ TAGS_METADATA = [
         "description": "Manage and obtain information about a particular project or PEP.",
     },
 ]
+
+INFO_KEY = "info"