+++ pephub/pepstat/const.py
@@ -1,0 +1,8 @@
+PKG_NAME = "pepstat"
+INDEX_CMD = "index"
+
+INDEX_STORE_KEY = "index_store"
+PROJECTS_KEY = "projects"
+INFO_KEY = "info"
+N_PROJECTS_KEY="n_projects"
+N_SAMPLES_KEY="n_samples"