+++ pephub/routers/version1.py
@@ -1,9 +1,10 @@
 import jinja2
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter
 from platform import python_version
 from starlette.requests import Request
 from starlette.templating import Jinja2Templates
 
+from .._version import __version__ as pephub_version
 from ..const import BASE_TEMPLATES_PATH
 from ..main import _PEP_STORES
 
@@ -20,6 +21,7 @@ router = APIRouter(
 )
 
 ALL_VERSIONS = {
+    "pephub_version": pephub_version,
     "peppy_version": peppy.__version__,
     "python_version": python_version(),
 }
@@ -28,9 +30,22 @@ ALL_VERSIONS = {
 @router.get("/")
 async def main(request: Request):
     templ_vars = {"request": request}
-    return templates.TemplateResponse("index.html", dict(templ_vars, **ALL_VERSIONS))
-
+    namespaces = _PEP_STORES.get_namespaces()
+    return templates.TemplateResponse(
+        "index.html", 
+        dict(
+            templ_vars, 
+            **ALL_VERSIONS,
+            namespaces=namespaces,
+        )
+    )
 
 @router.get("/pep-list")
 async def pep_list():
-    return _PEP_STORES
+    namespaces = _PEP_STORES.get_namespaces()
+    return [
+        dict(
+            **n, 
+            projects=_PEP_STORES.get_projects(n['name'])) 
+            for n in namespaces
+    ]